+++ src/main/java/controllers/MainMenu.java
@@ -1,6 +1,4 @@
-package Controllers;
-
-import UseCases.CreateGame;
+package controllers;
 
 public class MainMenu {
 //    The controller that leads to each use case
