+++ src/main/java/entities/Game.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 public class Game {
     /**
