+++ src/main/java/entities/GameStorage.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import javax.swing.*;
 import java.io.BufferedReader;
