+++ src/main/java/entities/Gunslinger.java
@@ -1,27 +1,27 @@
-package Entities;
+package entities;
 import Combat.Move;
 
 import java.util.ArrayList;
 
-public class Mage extends Player{
-    /* This class creates the mage class. The values of the class will be changed depending on the
-    * enemy the player encounters*/
+public class Gunslinger extends Player {
+    /* This class creates the Gunslinger class. The values of the class will be changed depending on the
+     * enemy the player encounters*/
     //TODO implement combat methods
 
-    public Mage(String name, int HP, int attackDamage, int damageMultiplier, int money, int XP
+    public Gunslinger(String name, int HP, int attackDamage, int damageMultiplier, int money, int XP
             , int max_XP, int player_level) {
         super(name);
     }
 
-    public Mage(String name) {
+    public Gunslinger(String name) {
         super(name);
     }
 
     @Override
     public ArrayList<Move> setMoves() {
-        Move first = new Move("Flame of the Fell God", 1, 1, "Opponent");
-        Move second = new Move("Comet Azur", 1, 1, "Opponent");
-        Move third = new Move("Stars of Ruin", 1, 1, "Opponent");
+        Move first = new Move("Go Go Glock", 1, 1, "Opponent");
+        Move second = new Move("gun-fu", 1, 1, "Opponent");
+        Move third = new Move("360 NoScope", 1, 1, "Opponent");
         Move fourth = new Move("Execution", 1, 1, "Opponent");
 
         ArrayList<Move> moves = new ArrayList<Move>();
@@ -32,17 +32,16 @@ public class Mage extends Player{
 
         return moves;
     }
-
     @Override
     public boolean attack(String move, Character characterBeingAttacked) {
 
         switch (move) {
-            case "Flame of the Fell God":
-                return flameOfTheFellGod(characterBeingAttacked);
-            case "Comet Azur":
-                return cometAzur(characterBeingAttacked);
-            case "Stars of Ruin":
-                return startsOfRuins(characterBeingAttacked);
+            case "Go Go Glock":
+                return gogoGlock(characterBeingAttacked);
+            case "gun-fu":
+                return gunfu(characterBeingAttacked);
+            case "360 NoScope":
+                return noScope(characterBeingAttacked);
             case "Execution":
                 return execution(characterBeingAttacked);
             default:
@@ -51,17 +50,17 @@ public class Mage extends Player{
         }
     }
 
-    private boolean flameOfTheFellGod(Character characterBeingAttacked) {
+    private boolean gogoGlock(Character characterBeingAttacked) {
         characterBeingAttacked.receiveDamage(2);
         return true;
     }
 
-    private boolean cometAzur(Character characterBeingAttacked) {
+    private boolean gunfu(Character characterBeingAttacked) {
         characterBeingAttacked.receiveDamage(2);
         return true;
     }
 
-    private boolean startsOfRuins(Character characterBeingAttacked) {
+    private boolean noScope(Character characterBeingAttacked) {
         characterBeingAttacked.receiveDamage(1);
         return true;
     }
