+++ src/test/java/PlayerTest.java
@@ -1,4 +1,4 @@
-import Entities.Player;
+import entities.Player;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
